feat!: move stuff to utils package
