+++ base/src/commonMain/kotlin/com/ekezet/hurok/utils/LoopExt.kt
@@ -1,10 +1,13 @@
-package com.ekezet.hurok
+package com.ekezet.hurok.utils
 
+import com.ekezet.hurok.Action
+import com.ekezet.hurok.Loop
+import com.ekezet.hurok.ViewState
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
 
 /**
- * The first state of this [Loop] in a blocking fashion. Useful for testing.
+ * The first state of this [com.ekezet.hurok.Loop] in a blocking fashion. Useful for testing.
  */
 val <TState : ViewState<TModel, TDependency>, TModel : Any, TArgs : TModel, TDependency, TAction : Action<TModel, TDependency>> Loop<TState, TModel, TArgs, TDependency, TAction>.firstState: TState
     get() = runBlocking { state.first() }
