+++ base/src/commonMain/kotlin/com/ekezet/hurok/utils/Tools.kt
@@ -1,12 +1,17 @@
-package com.ekezet.hurok
+package com.ekezet.hurok.utils
+
+import com.ekezet.hurok.Action
+import com.ekezet.hurok.Loop
+import com.ekezet.hurok.Renderer
+import com.ekezet.hurok.ViewState
 
 /**
- * Render a given [Loop] into a state.
+ * Render a given [com.ekezet.hurok.Loop] into a state.
  *
- * @param constructor a [Loop] constructor reference
+ * @param constructor a [com.ekezet.hurok.Loop] constructor reference
  * @param model model used for rendering
- * @param renderer [Renderer] instance used for rendering
- * @param args optional arguments for the [Loop]
+ * @param renderer [com.ekezet.hurok.Renderer] instance used for rendering
+ * @param args optional arguments for the [com.ekezet.hurok.Loop]
  * @return a state based on the input arguments
  */
 inline fun <TState : ViewState<TModel, TDependency>, TModel : Any, TArgs, TDependency, TAction : Action<TModel, TDependency>, TRenderer : Renderer<TModel, TState>, TLoop : Loop<TState, TModel, TArgs, TDependency, TAction>> renderState(
