+++ base/src/jvmTest/kotlin/com/ekezet/hurok/RenderStateTest.kt
@@ -3,6 +3,7 @@ package com.ekezet.hurok
 import com.ekezet.hurok.test.TestLoop
 import com.ekezet.hurok.test.TestModel
 import com.ekezet.hurok.test.TestState
+import com.ekezet.hurok.utils.renderState
 import kotlin.test.Test
 import kotlin.test.assertEquals
