+++ compose/src/androidMain/kotlin/com/ekezet/hurok/compose/LoopExt.android.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.ekezet.hurok.Action
 import com.ekezet.hurok.Loop
 import com.ekezet.hurok.ViewState
-import com.ekezet.hurok.firstState
+import com.ekezet.hurok.utils.firstState
 import com.ekezet.hurok.test.CoverageIgnore
 
 /**
