+++ compose/src/commonMain/kotlin/com/ekezet/hurok/compose/LoopExt.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.collectAsState
 import com.ekezet.hurok.Action
 import com.ekezet.hurok.Loop
 import com.ekezet.hurok.ViewState
-import com.ekezet.hurok.firstState
+import com.ekezet.hurok.utils.firstState
 import com.ekezet.hurok.test.CoverageIgnore
 
 /**
